Path Resolution Mode SASServer/SASWorkstation and encoding
Hy Tim,
First, I want to say  that these works are really useful for teaching Actuarial Sciences in SAS.
I'm a actuary and teaching at Le Mans University (France).
My GitHub push aims to solve two issues :
First, the encoding of ouput: utf8 is a better standart than ANSI for langage that use accentuated character (I'm french). I mainly handled this issue with filenames using the encoding option. By default, it is utf8, but it can be changed by a new custom variable or deduced from the encoding option in the inputenc package.

The second issue is the current path. You solved it in a super simple and effective way with ./.  Howerver on Viya for Learners/Educators, I receive the following log message:

80   %include "./myfile.sas";
WARNING: Physical file does not exist,
         /opt/sas/viya/config/var/tmp/compsrv/default/08273935-f335-49d5-b522-7415a65871c3/SAS_workF53C000001F5_sas-compute-server-8
         934936c-0b94-4d1c-8a57-fc1b6768a86b-12955/myfile.sas.

It appears that the package references a computation directory rather than the directory where my file is located. To address this issue, I have modified the .sty file by adding a SASmode option. I would be happy to share this update with you.

I would be happy to share this update with you.

Best regards,
Martial